Added javadocs

+++ src/com/iDocent/AccessPointDB.java
@@ -12,6 +12,10 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * This is an object to hold the list of dowloaded APs and their location
+ *
+ */
 //A database that will map an ssid to an xyz location
 public class AccessPointDB {
 	private HashMap<String, LinkedList<Float>> accessPoints;
@@ -23,7 +27,6 @@ public class AccessPointDB {
 //	private static final int z=2; 
 //	
 //	private iDocent miD;
-	
 	public AccessPointDB(iDocent miD)
 	{
 		//this.miD = miD;
@@ -39,7 +42,7 @@ public class AccessPointDB {
 	                         
 	@param  mac - the MAC address of the desired access point
 	
-	@return List<Integer> - the access point's location stored in a list (x y z)
+	@return List<Float> - the access point's location stored in a list (x y z)
 	 *   
 	 */
 	public List<Float> GetLocation(String mac)
@@ -47,6 +50,11 @@ public class AccessPointDB {
 		return accessPoints.get(mac.toLowerCase());
 	}
 
+	/**
+	 * Start scanning APs.  Checks if there is a live connection with the server then downloads the list
+	 * of APs.
+	 * @return boolean - true if the list was downloaded
+	 */
 	public boolean StartScanLoop() {		
 		//Download the list of access points from the server
 		if(!irl.isConnected())
@@ -59,10 +67,17 @@ public class AccessPointDB {
 		return true;
 	}
 
+	/**
+	 * Disconnect from the server.
+	 */
 	public void EndScanLoop() {
 		irl.Disconnect();
 	}
 
+	/**
+	 * Connect to the server
+	 * @return boolean - true if connection was successful.
+	 */
 	public boolean Connect() {
 		irl.Connect();
 		return irl.isConnected();
